[KS-APP-02] StockAdapter 클래스를 별도 파일로 만들어 관리하자
